Fix IterateERC20ToDenom
The function IterateERC20ToDenom is only used once, in genesis export to
export the entire mapping.

In the recent commit "Use geth address struct instead of strings
internally" this function was not updated to parse the updated save
format.

This patch corrects the issue and adds a test to the existing migration
integration tests.