Keep airdrop integration test from stalling
In order to pack as many recipients as possible into the airdrop
proposal transaction each users amount is limited to u64 max, while this
would be impratical for Ethereum based tokens it's completely reasonable
for CosmosSDK tokens with 6 decimals of precision.

As the final step of the airdrop proposal integration test we would try
to generate an airdrop proposal which spends more than the tokens
remaining in the community pool. This would sometimes cause the test to
lockup trying to generate an invalid proposal when it was simply not
possible.

If the amount of tokens remaining in the community pool is greater than
the maximum aridrop size times u64 max the function
generate_invalid_accounts_and_amounts would continue to run forever.

This commit adds a few unit tests that ensure convergence in the future
and detects the impossible case and skips the last part of the test