Modify whitelist relaying to take a price
This patch modifies the whitelisted token relaying functionality to take
an oracle price via the config rather than a simple threshold. The
previous method, where any amount of a token over a fixed amount was
enough to relay, did not actually approximate profitability in any way.
A 100tx batch with fee fee value x would be considered just as
profitable as a 1tx batch with the same fee amount but a much higher gas
cost.