Eliminate IBC relayer start mode for test runner
Previously there was a separate run mode of the integration test runner
that would only start the IBC relayer and then run in the background
continuously. This caused a lot of race conditions because the
background relayer was using the first validators private key to send
ibc updates.

This key is the obvious one to use and so it's used in many places to
send messages, resulting in a sequence fight between the Hermes relayer
and the test runner.

This patch moves the setup of Hermes into the main
run to help coordinate and prevent sequence issues, only a single send
is sent from keys[0] to a special relayer key, which will then send all
future background messages and avoid sequence fighting with any other
part of the test runner.