Expand altruistic relayer samples
As it turns out with a slow relayer loop time, like 5 minutes a days
worth of samples is not nearly enough relay only once per day. This
patch is not a very well thought out compentation for relayer loop time,
but instead just expands the number of samlpes to a much higher value.

A proper fix would take into account loop time and a better
understanding of how and when samples are pushed from the cache.