Bump for v1.5.2 Mercury2.0 upgrade
This release contains fixes for the IBC auto-forwarding tools as well as
some updates to gbt.

The previous v1.5.1 release did not have an actual upgrade handler.
Since this is otherwise a consensus compatible upgrade it's the upgrade
handler itself that breaks backwards compatibility here.

We've added the upgrade handler to try and fit the upgrades into a
standard format.