Exclude test-runner from automated release builds
Since the IBC test was introduced test-runner uses unix stdio for
handling input/output with Hermes being run as a subprocess. This
obviously doesn't build on Windows and may not build on Mac. This patch
simply doesn't build test-runner for the automated release since we
would just throw away the artifact anyways.