Bump for gbt v1.5.3
This release contains changes *only* for the standalone relayer command
`gbt relayer`. There was a bug that prevented the standalone relayer
(but not the orchestrator integrated relayer) from relaying messages.
This patch is purely to fix that bug and contains no changes to either
the orchestrator or the gravity binary.