Always boost gas estimates by 20%
Due to the way eip1559 works block gas prices may go up or down by as
much as 12.5% every block. Since we're trying to get our transaction
into the next block we're effectively betting that the gas price is
trending level or down when we submit a tx with the last blocks base fee
attached.

This assumption was observed to be wrong too often for practical usage.
With submissions constantly failing due to an upward trending gas price.
This patch reduces the issue significantly by baking a 20% fudge factor
into all estimated costs, with the assumption that our tx will get into
the next block.

This assumption is itself still flawed, block times are unpredictable
and we don't know when in the cycle we are making our prediction, we
could be reading the base fee for block n-2 miss block n-1 and need to
correctly predict the base fee in block n. A factor of 20% is too little
to cover the worst case 25% increase over two blocks.

Ultimately we'll see if this is a useful heuristic.