Tool: Add manual upgrade tester
Usage: With the chain at the block you want to test using
the old binary, apply this commit and modify it for your upgrade.
In particular: Modify SetupManualTest() and RunManualTestingTool()
and then build a binary. Run this new binary with
GRAVITY_UPGRADE_HEIGHT=<CURRENT HEIGHT + 1> gravity start ...

Update: Configured for Neutrino upgrade