Attestation checks in upgrade test, EthereumClaim in Rust
Surrounding an upgrade it is possible to lose or misplace data, and not
know until deploying to mainnet and trying to use that missing data. By
adding a check for attestations around the upgrade integration test we
can assert that none of the test attestations are missing after the
upgrade.

EthereumClaim has been reproduced in gravity_proto but without the
ClaimHash() function