Limit oracle submissions
As an optimization strategy the orchestrator bundles as many oracle
submissions as possible into a single message. Normally this works
great but recently a validator that has not been active for some time
unajailed nad has found their orchestrator unable to catch up because it
can not submit the thousands of events as a single message.

This patch adds a constant cap to the number of oracle events in a
single transaction which would prevent this scenario.

Making this patch work properly reveals the need for trait based events
again, this code is just generally awkward. Last time this was attempted
trait objects where problematic, but perhaps things have moved forward
since then.