Remove merge countdown, modify block delay
This patch *is not* a complete ethereum proof of stake update for the
Orchestrator. It only removes the merge countdown and sets a 64 block
delay for assumed finality.

Previously we waited 13 blocks for finality given known probabilities of
proof of work re-orgs but there was no way to really 'know' when
somthing was final.

ETH2 does provide actual finality. The problem is it takes a while and
is difficult to query. Mapping consensus layer slots to execution layer
blocks is a complex process best left to an API query which does not
seem to exist, at least in any obvious form.

This patch uses a 96 block delay which is in all probability safer than
our previous proof of work delay, but does not check for actual ETH2
finality, we will update GBT to do this in the future once we have a
100% reliable way to map slots/epochs to execution layer blocks.

We could attempt a hardcoded offset mapping, but in the case where many
slots pass with no block block delay finality is safer than using a
hardcoded offset to map execution layer blocks to slots since empty
slots will shrink the rolling window and silently reduce saftey whereas
a block delay at the execution layer becomes more likley to be finlaized
if slots are missed at the consensus layer.