Fixup denom-metadata preflight checks
This patch fixes a logic issue with the denom-metadata checks,
specifically the code was well prepared for a specific allowed edge
case, changing the metadata of a token before an erc20 representation
has been deployed.

But in trying to support this edge case the code ran into several issues
around tokens that have not yet had any metadata deployed, namely that
the queries used returned nil and this went unchecked.

This patch modifies the behavior to not support modifying metadata
denoms which already have it set even though this is just client side
verification and the tx may actually be accepted by the chain.

The primary reason for this reduction in functionality is that we need
exceptionally clear error messages for this functionality since
submitters are often confused and using ibc hashes from for example the
wrong chain. On the other hand we've never seen anyone actually try to
modify the metadata before deploying they usually only notice issues
after deploying the erc20 on Ethereum.