Add Cross bridge balance checking, eth balances in claims
In order to ensure bridge security, orchestrators will be required to
submit some of the ERC20 balances of Gravity.sol, based on a governance
vote. The community will vote on which tokens are of significant
importance to the bridge (probably based on TVL) and orchestrators will
read the parameter set by the vote. Then they will automatically fetch
historical balances of Gravity.sol for each of those ERC20s for each of
their Ethereum event claims (Valset updates, ERC20 deploys, Batch
executions, Sends to Cosmos, Logic call executions). The cosmos side of
the bridge will check the balances against what the orchestrators have
attested to, halting the bridge if necessary.