Add MsgSendToEth chain fee
In proposal #86 the Gravity Bridge community voted to include a minimum
fee on Sends to Ethereum which will be paid out to Gravity Bridge
stakers and validators. This is the first part of that development where
we add the ChainFee field to the MsgSendToEth proto, but collecting and
distributing fees will be in another commit.