Enforce 10% minimum comission
As outlined in proposal #106 which has now passed minimum comission
should be set to 10% on the next upgrade.

There are two parts to the minimum comission. The transaction
ante-handler and the upgrade handler. The upgrade handler edits
validators as the update occurs to enforce the minimum comission.

The ante-handler prevents the creation or editing of a validator to have
a comission below the limit.

This patch just integrates the Mercury upgrade handler with the new
constant and updates the existing ante-hanlder with the new constant.