Fix invariant valset power check
As it turns out this invariant is run before some of the slashing
handling in endblocker order. With normal delegating and unbonding this
check will function fine but a significant portion of the validator set
being slashed will trigger it.

This patch simply removes the check as there's no easy way to keep it
wihtout risky adjustments to the endblocker order.