Fix automated upgrade test
This fixes the automated upgrade test by downloading the test_runner
binary itself to run the upgrade test v1 from that release. This does
add the unexpected complexity of running the old version of part 1 of
the test, which might not line up with the new part 2. But it does get
things running.

In the long term the best solution to this problem is to break the test
runner into a library so that it can be imported as a rust dependency.