Add chain fee support for Cosmos to Eth and Recover Funds
Since the chain fee is automatically deducted from users' accounts, this
commit requires an explicit chain fee argument for Cosmos to Eth. For
the Recover Funds command --send-to-eth case the chain fee arg is omissible but a
reasonable value will be deducted from the --amount provided.

As users are likely to bridge ALL of the funds when using Recover Funds,
this default functionality is desireable. For the Cosmos to Eth command
users should be aware of the full fees impacting them before bridging.