Move fee deduction from AnteHandler to MsgServer
Fees collected in an AnteHandler will be collected regardless of Msg
success or failure. A typo in a MsgSendToEth could result in significant
fees and frustration for users, and doesn't really accomplish anything.
Spam prevention is handled by the validator minimum fee selection, so by
moving the fee deduction to msg_server.go users receive a better
experience on mistakes.