Upgrade to cosmos-sdk v0.45.13, althea-net/bech32-ibc
It appears that osmosis-labs/bech32-ibc is no longer being supported
given that osmosis-labs/osmosis no longer imports it and no new releases
have been created, so althea-net/bech32-ibc v0.4.2 is the version we are
using from here on out.

Additionally the sdk version and ethermint version have been locked in
the go.mod with replace directives, these must be deleted to update
these dependencies in the future.