Remove dragonberry replace directive
github.com/confio/ics23/go was initially replaced with a fixed version
hosted at github.com/cosmos/cosmos-sdk/ics23/go in an emergency fix for
the dragonberry vulnerability. Since new sdk releases have come out, an
official release of confio/ics23/go v0.9.0 was made with the fixes for
dragonberry, and also the repo was rehosted at github.com/cosmos/ics23/go.

Due to go weirdness we still have to import the confio repo and not the
cosmos repo, but nevertheless we are up to date for dragonberry and not
locked out of updates with this replace directive removed.