Implement Cross Bridge Balance checking in gravity
Many bridges attempt to increase security by placing rate limits on
token transfers. These limits are met when a large opportunity, an
uptick in popularity, or some black swan event occurs. Therefore rate
limits are an anti-user attempt at achieving stronger security. Rather
than go with this approach, we are implementing cross bridge balance
checking on two fronts. This commit contains the first front of checking
and setup for the second.

On the Cosmos side the gravity module performs an out-of-band
check on the state changes each observed Attestation makes, requiring
that precisely the expected bank supply / gravity module balance changes
occur separate from the actual state transition logic.

To set up for the Orchestrator side balance checking, the gravity module
publishes "Bridge Balance Snapshots", which are records produced on each
attestation application and contain the following:
* The x/bank supply of any Ethereum-originated asset
* The gravity module balance of any bridged Cosmos-originated asset
* The cosmos block height on which the attestation was applied
* The ethereum block height of the event which generated this
  attestation
Orchestrators will be required to monitor these "Bridge Balance
Snapshots" via GRPC and check the historical ERC20 balances of the
Gravity.sol contract at each snapshot's ethereum block height. If an
unexplained discrepancy is detected, Orchestrators are required to halt
the submission of claims and Oracles are required to halt the submission
of signatures to Cosmos. This will effectively halt the bridge until
either an ethereum fork is resolved or the bridge is manually attended
to.