Add balance checking to Orchestrator and Oracle
This commit implements the second front of bridge balance checks, where
the Orchestrator + Oracle monitor both sides of the bridge and
enter a "holding pattern" when a discrepancy is detected.

A new function has been added which will be called each iteration by the
Orchestrator and the Oracle. This function queries the Cosmos side for
the tokens to monitor, then checks the Ethereum side balances for each
snapshot stored on the Cosmos side. If all that checks out, the
iteration progresses as normal. If the Ethereum side is lower than the
Cosmos side (indicating a theft) then the iteration ends early and
monitoring starts again next iteration. Notably the Ethereum side can
become inflated by sending an ERC20 the Gravity.sol contract (instead of
calling .sendToCosmos()), so the Ethereum-side checks only assert that
the Gravity.sol balance is enough to handle any withdrawals.

If the balance discrepancy is due to an Ethereum fork, the "holding
pattern" will ensure that bridge operators resume once the fork is
resolved.