Reenable the upgrade test
This will work although it won't be running the correct orchestrator.

This test will not work when backwards incompatible changes to the orch
are made as we can't run the new version of the orch included in the
test runner against the old version of the chain. But for this upgrade
that is not an issue.