Set default grpc and eth addresses for client ops
This patch sets user-friendly default rpc addresses for client
operations. For the relayer and orchestrator the use of local nodes is
important to ensure reliability. If too many orchestrators or relayers
became dependent on a single node set by using the defaults it could
impact the bridge.

But for simple one off queries or commands the user can notice the
server is down and substitute any one of many public servers easily.
Meaning we can safely default to more user friendly behavior.