Use proper valset, signatures lookup for spot relay
Previously the spot relay command used a recent valset simply for ease.
But as it turns out it's not that complicated at all to lookup the
correct valset.

Furthermore we where trying to query signatures we looked up the wrong
ones, using the gravity address rather than the target token address