Fix chain setup for new auction module params
One problem the auction module introduces is the need to use ugraviton
as the staking token, so that we can validate in CI that the auction
module will never allow auctions of the ugraviton balance in the
community pool.

Additionally, the default params values for the auction module were
chosen with mainnet in mind, but waiting 7 days for auctions to complete
makes for an infeasible test. More sensible defaults have been chosen
for the test environment.