Add cargo ACL
THis is a project to help make Rust supply chain attacks harder by
specifying specific permissions for dependencies and allowing them
individually.

https://davidlattimore.github.io/making-supply-chain-attacks-harder.html

I'm impressed by how easy the setup process was and while it doesn't
verify that no supply chain attacks exist, and is possible to circumvent
it will at least prevent the silent addition of a permission grabbing
dependency.

Co-authored-by: Christian Borst <christian@althea.systems>