Add JSONRPC Server for MetaMask transaction signing
The Gravity JSONRPC Server is a mock Ethereum JSONRPC server which
returns largely incorrect responses as a minimum effort to enable
MetaMask signing. This server performs the essential function of
tricking MetaMask into thinking that Gravity is a real EVM chain with
the standard EVM JSONRPC interface so that it will allow signatures to
be signed with EIP-712 for Gravity Bridge Chain.

Without this server, MetaMask fails several validation steps required
before signing any transactions.