Correct signature maleability
This patch corrects a signature maleability vulnerability that would
allow a validator to potential perform a denial of service attack on the
chain. A validator could submit a maelabile mirror signature and have it
accepted into the chain, but it would not be accepted as valid by either
the relayer or Ethereum itself.

This would result in a halt of some relayers until a fix to skip the
signature was releasesd.

This vulnerability was reported by https://twitter.com/dooflin5 and ging3r