Reduce get_attestations() limit
This patch reduces the limit in the get_attestations function, while
this function is only used in tests in the main gravity repo it is used
by the gravity info server to determine when transactions are being
finalized.

The 1000 attestation limit is too large and runs into response size
errors.