Optimize relayer gas calldata usage
When Ethereum is charging for calldata it provides a discount for zero
bytes versus nonzero bytes. Zero bytes are only charged 4 gas per bytes,
whereas nonzero bytes are charged 16 gas per byte.

This steep difference sets up this optimization. Gravity.sol is designed
to allow zeroed out signatures to deal with validators who did not
submit a signature for a given item.

This means we can choose to zero out signatures for validators that
aren't required to execute a batch or validator set update. Paying only
the lower cost for zeroed calldata.

The way this is implemented is by adding an optimize flag to the
signature ordering step allowing us to choose an optimized signature set
or not. Optimized signature submission is disabled on validator sets
currently out of an abundance of caution. Not becuase zeroed signatures
present any security issue, but because they may cause a liveness issue
with submitting messages if the threshold is miscomputed in some way and
not being able to update the validator set would be a fairly critical
liveness issue.