Add gravity v2 PendingSendToEthV2[BySender]
PendingSendToEthV2 is an updated version of the PendingSendToEth GRPC
query endpoint which returns the same information, but for the batched
transfers the batch and transaction are given. There is additionally the
PendingSendToEthV2BySender endpoint which enables filtering by the
transaction sender.

This has been implemented in a new v2 namespace which will house the newer
queries and types for Gravity.